### 2023-07-03 #5613 @editorialbot
[REVIEW]: FuzzyClass: An R package for Fuzzy and Non-Fuzzy probabilistic-based classifiers
**Submitting author:** @Jodavid (<a href="http://orcid.org/0000-0002-2131-6464">Jodavid Ferreira</a>)
**Repository:** https://github.com/leapigufpb/FuzzyClass/
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.5
