+++ R/FuzzyBayesRule.R
@@ -1,6 +1,6 @@
-#' Fuzzy Gamma Naive Bayes
+#' Fuzzy Bayes Rule
 #'
-#' \code{FuzzyGammaNaiveBayes} Fuzzy Gamma Naive Bayes
+#' \code{FuzzyBayesRule} Fuzzy Bayes Rule
 #'
 #'
 #' @param train matrix or data frame of training set cases.
@@ -11,7 +11,7 @@
 #' @return A vector of classifications
 #'
 #' @references
-#' \insertRef{de2018fuzzy}{FuzzyClass}
+#' \insertRef{roneide2006fuzzy}{FuzzyClass}
 #'
 #' @examples
 #'
@@ -26,7 +26,7 @@
 #' # matrix or data frame of test set cases.
 #' # A vector will be interpreted as a row vector for a single case.
 #' test <- Test[, -5]
-#' fit_NBT <- FuzzyGammaNaiveBayes(
+#' fit_NBT <- FuzzyBayesRule(
 #'   train = Train[, -5],
 #'   cl = Train[, 5], cores = 2
 #' )
@@ -38,28 +38,26 @@
 #' @importFrom stats dgamma
 #'
 #' @export
-FuzzyGammaNaiveBayes <- function(train, cl, cores = 2, fuzzy = TRUE) {
-  UseMethod("FuzzyGammaNaiveBayes")
+FuzzyBayesRule <- function(train, cl, cores = 2, fuzzy = TRUE) {
+  UseMethod("FuzzyBayesRule")
 }
 
 #' @export
-FuzzyGammaNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = T) {
+FuzzyBayesRule.default <- function(train, cl, cores = 2, fuzzy = T) {
 
   # --------------------------------------------------------
   # Estimating class parameters
-  train <- as.data.frame(train)
-  cols <- ncol(train) # Number of variables
-  if(is.null(cols)){
-    cols <- 1
-  }
-  dados <- train # training data matrix
-  M <- c(unlist(cl)) # true classes
-  M <- factor(M, labels = sort(unique(M)))
+  p_data <- predata(train,cl)
+  # --
+  train <-  p_data$train
+  cols <- p_data$cols
+  dados <- p_data$dados
+  M <- p_data$M
   # --------------------------------------------------------
 
   # --------------------------------------------------------
-  # Estimating Gamma Parameters
-  parametersC <- estimation_parameters_gamma(M, cols, dados)
+  # Estimating Parameters
+  parametersC <- estimation_parameters_fuzzy_bayes_rule(M, cols, dados)
 
   # --------------------------------------------------------
   Sturges <- Sturges(dados, M);
@@ -86,21 +84,21 @@ FuzzyGammaNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = T) {
     pk = pk,
     fuzzy = fuzzy
   ),
-  class = "FuzzyGammaNaiveBayes"
+  class = "FuzzyBayesRule"
   )
 }
 # -------------------------
 
 
 #' @export
-print.FuzzyGammaNaiveBayes <- function(x, ...) {
+print.FuzzyBayesRule <- function(x, ...) {
   if (x$fuzzy == T) {
     # -----------------
-    cat("\nFuzzy Gamma Naive Bayes Classifier for Discrete Predictors\n\n")
+    cat("\nFuzzy Bayes Rule Classifier for Discrete Predictors\n\n")
     # -----------------
   } else {
     # -----------------
-    cat("\nNaive Gamma  Bayes Classifier for Discrete Predictors\n\n")
[…]
