+++ R/FuzzyGeoNaiveBayes.R
@@ -1,0 +1,301 @@
+#' Fuzzy Naive Bayes Geometric Classifier
+#'
+#' \code{FuzzyGeoNaiveBayes} Naive Bayes Geometric Classifier
+#'
+#'
+#' @param train matrix or data frame of training set cases.
+#' @param cl factor of true classifications of training set
+#' @param cores  how many cores of the computer do you want to use (default = 2)
+#' @param fuzzy boolean variable to use the membership function
+#'
+#' @return A vector of classifications
+#'
+#' @references
+#' \insertRef{moraes2021new}{FuzzyClass}
+#'
+#' @examples
+#'
+#' set.seed(1) # determining a seed
+#' class1 <- data.frame(vari1 = rgeom(100,prob = 0.2),
+#'                      vari2 = rgeom(100,prob = 0.2),
+#'                      vari3 = rgeom(100,prob = 0.2), class = 1)
+#' class2 <- data.frame(vari1 = rgeom(100,prob = 0.5),
+#'                      vari2 = rgeom(100,prob = 0.5),
+#'                      vari3 = rgeom(100,prob = 0.5), class = 2)
+#' class3 <- data.frame(vari1 = rgeom(100,prob = 0.9),
+#'                      vari2 = rgeom(100,prob = 0.9),
+#'                      vari3 = rgeom(100,prob = 0.9), class = 3)
+#' data <- rbind(class1,class2,class3)
+#'
+#' # Splitting into Training and Testing
+#' split <- caTools::sample.split(t(data[, 1]), SplitRatio = 0.7)
+#' Train <- subset(data, split == "TRUE")
+#' Test <- subset(data, split == "FALSE")
+#' # ----------------
+#' # matrix or data frame of test set cases.
+#' # A vector will be interpreted as a row vector for a single case.
+#' test <- Test[, -4]
+#' fit_NBT <- FuzzyGeoNaiveBayes(
+#'   train = Train[, -4],
+#'   cl = Train[, 4], cores = 2
+#' )
+#'
+#' pred_NBT <- predict(fit_NBT, test)
+#'
+#' head(pred_NBT)
+#' head(Test[, 4])
+#' @export
+FuzzyGeoNaiveBayes <- function(train, cl, cores = 2, fuzzy = T) {
+  UseMethod("FuzzyGeoNaiveBayes")
+}
+
+#' @export
+FuzzyGeoNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = T) {
+
+  # --------------------------------------------------------
+  # Estimating class parameters
+  p_data <- predata(train,cl)
+  # --
+  train <-  p_data$train
+  cols <- p_data$cols
+  dados <- p_data$dados
+  M <- p_data$M
+  intervalos <- p_data$intervalos
+  #--------------------------------------------------------
+
+
+  #------------------------------------------------------------
+  # Verify data types
+  verifyNumbers <- sapply(1:cols, function(i){
+    set.seed(3)
+    n = 3
+    subset <- sample(dados[,i],size = n, replace = F)
+    result <- subset == floor(subset)
+    if(sum(result) == n){
+      result <- 1
+    }else{
+      result <- 0
+    }
+    return(result)
+  })
+
+  # --------------------------------------------------------
+  if(sum(verifyNumbers) != cols){ stop("All variables must be discrete values.") }
+
+  # --------------------------------------------------------
+  # --------------------------------------------------------
+  # Estimating class memberships
+  pertinicesC <- lapply(1:length(unique(M)), function(i) {
+    lapply(1:cols, function(j) {
+      SubSet <- dados[M == unique(M)[i], j]
+      getMemberships(SubSet, intervalos)
+    })
+  })
+  # --------------------------------------------------------
+  # --------------------------------------------------------
+  # Estimating Geometric Parameters
[…]
