+++ inst/REFERENCES.bib
@@ -1,3 +1,31 @@
+@article{bian2015sar,
+  title={SAR probability density function estimation using a generalized form of K-distribution},
+  author={Bian, Yong and Mercer, Bryan},
+  journal={IEEE Transactions on Aerospace and Electronic Systems},
+  volume={51},
+  number={2},
+  pages={1136--1146},
+  year={2015},
+  publisher={IEEE}
+}
+
+@article{roneide2006fuzzy,
+  title={A Fuzzy Bayes Evaluator for On-Line Training Evaluation Based on Virtual Reality},
+  author={de Moraes, Ronei Marcos and dos Santos Machado, Liliane},
+  year={2006}
+}
+
+
+@article{marcos2020double,
+  title={A double weighted fuzzy gamma naive bayes classifier},
+  author={Moraes, RM and Soares, EAMG and Machado, LS},
+  journal={Journal of Intelligent \& Fuzzy Systems},
+  volume={38},
+  number={1},
+  pages={577--588},
+  year={2020},
+  publisher={IOS Press}
+}
 
 @inproceedings{moraes2009another,
   title={Another approach for fuzzy naive bayes applied on online training assessment in virtual reality simulators},
@@ -108,4 +136,15 @@
   volume={1},
   pages={57--68},
   year={2016}
+}
+
+@article{frery1997model,
+  title={A model for extremely heterogeneous clutter},
+  author={Frery, Alejandro C{\'e}sar and Muller, H-J and Yanasse, Corina da Costa Freitas and Sant'Anna, Sidnei Jo{\~a}o Siqueira},
+  journal={IEEE transactions on geoscience and remote sensing},
+  volume={35},
+  number={3},
+  pages={648--659},
+  year={1997},
+  publisher={IEEE}
 }
