+++ R/FuzzyTrapeNaiveBayes.R
@@ -47,16 +47,14 @@ FuzzyTrapezoidalNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = T) 
 
   # --------------------------------------------------------
   # Estimating class parameters
-  train <- as.data.frame(train)
-  cols <- ncol(train) # Number of variables
-  if(is.null(cols)){
-    cols <- 1
-  }
-  dados <- train # training data matrix
-  M <- c(unlist(cl)) # true classes
-  M <- factor(M, labels = sort(unique(M)))
-  intervalos <- 10 # Division to memberships
-  # --------------------------------------------------------
+  p_data <- predata(train,cl)
+  # --
+  train <-  p_data$train
+  cols <- p_data$cols
+  dados <- p_data$dados
+  M <- p_data$M
+  intervalos <- p_data$intervalos
+  #--------------------------------------------------------
   # --------------------------------------------------------
   # Estimating class memberships
   pertinicesC <- lapply(1:length(unique(M)), function(i) {
