+++ R/FuzzyBinomialNaiveBayes.R
@@ -61,17 +61,16 @@ funcao_estimation_N <- function(n, x){
 #' @export
 FuzzyBinomialNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = T) {
 
-  #--------------------------------------------------------
+  # --------------------------------------------------------
   # Estimating class parameters
-  train <- as.data.frame(train)
-  cols <- ncol(train) # Number of variables
-  if(is.null(cols)){
-    cols <- 1
-  }
-  dados <- train # training data matrix
-  M <- c(unlist(cl)) # true classes
-  M <- factor(M, labels = sort(unique(M)))
-  #--------------------------------------------------------
+  p_data <- predata(train,cl)
+  # --
+  train <-  p_data$train
+  cols <- p_data$cols
+  dados <- p_data$dados
+  M <- p_data$M
+  intervalos <- p_data$intervalos
+  # --------------------------------------------------------
   # --------------------------------------------------------
   # Verify data types
   verifyNumbers <- sapply(1:cols, function(i){
