+++ R/FuzzyBetaNaiveBayes.R
@@ -45,16 +45,15 @@ FuzzyBetaNaiveBayes <- function(train, cl, cores = 2, fuzzy = TRUE) {
 #' @export
 FuzzyBetaNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = TRUE) {
 
-  #--------------------------------------------------------
+  # --------------------------------------------------------
   # Estimating class parameters
-  train <- as.data.frame(train)
-  cols <- ncol(train) # Number of variables
-  if(is.null(cols)){
-    cols <- 1
-  }
-  dados <- train # training data matrix
-  M <- c(unlist(cl)) # true classes
-  M <- factor(M, labels = sort(unique(M)))
+  p_data <- predata(train,cl)
+  # --
+  train <-  p_data$train
+  cols <- p_data$cols
+  dados <- p_data$dados
+  M <- p_data$M
+  intervalos <- p_data$intervalos
   #--------------------------------------------------------
   for(j in 1:cols){
     dados[, j] <- dados[, j] / (max(dados[, j])+1e-2)
