### 2023-08-04 #5613 @MikeLydeamore
Hi @ajstewartlang @tianyi93. I have completed my review of this work. The paper itself is well written, and bar one typo (exponencial distribution), the only thing missing is a statement of other software solutions that exist for this problem.

I believe there are some issues with the codebase and repository which should be amended before publication:

- [ ] Please fix the `pkgdown` site: https://leapigufpb.github.io/FuzzyClass/reference/index.html gives me a 404 error
- [ ] Some automated testing of the software (beyond the current spellcheck) would be a great addition to robustness.
- [ ] Please add at least one vignette. This is a requirement for CRAN these days.
- [ ] Please add contribution guidelines to the README
- [ ] Please add issue reporting guidelines to the README
- [ ] The example in the README could do with much stronger documentation/commenting. It runs successfully but I don't know what it's supposed to do, or what the outputs mean.
- [ ] Dependencies: The `caret` library is used in the example, but this isn't included in the package install. It's a massive library, so I understand not wanting to include it as a dependency, but either an alternative `confusionMatrix` function should be sourced, or it needs to go into `DEPENDS`.

Finally, the example uses a function `GauNBFuzzyParam`. The documentation for this has a few issues:
- [ ] Parameter `w` documentations talks about `metd == 4`, but the documentation about `metd` only says this can range between 1 and 3
- [ ] Argument `cores` isn't used in this function (passed through elsewhere), and this should be stated as I was looking for the parallel pool to spawn in this function, not elsewhere.
