### 2023-08-06 #5613 @tianyi93
I have completed my review of this work. Few issues would suggest to amend before publication:

>Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

- [ ] Readme: please include running the command from R console

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

- [ ] No clear guidelines for third parties wishing to contribute to the software or seek support

>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

- [ ] Did not see much  core functionality of the software documented

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?`

- [ ] Please add the above

>State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [ ] there is no explicit description of how the "FuzzyClass" software compares to other commonly-used packages for classification tasks. The text primarily focuses on introducing the concept of Naive Bayes and Fuzzy Naive Bayes classifiers, their probability calculations, and the motivating example using the "FuzzyClass" package with real data.
