### 2023-08-10 77649db
update FuzzyClass
