+++ R/ExpNBFuzzyParam.R
@@ -7,9 +7,7 @@
 #' @param cl factor of true classifications of training set
 #' @param alphacut value of the alpha-cut parameter, this value is between 0 and 1.
 #' @param metd Method of transforming the triangle into scalar, It is the type of data entry for the test sample, use metd 1 if you want to use the Yager technique, metd 2 if you want to use the Q technique of the uniformity test (article: Directional Statistics and Shape analysis), and metd 3 if you want to use the Thorani technique
-#' @param alp When metd for 4, it is necessary to have alp which are alpha-cut defined
-#' @param w When metd for 4, it is necessary to have w which are alpha-cut weights defined
-#' @param cores  how many cores of the computer do you want to use (default = 2)
+#' @param cores  how many cores of the computer do you want to use to use for prediction (default = 2)
 #'
 #' @return A vector of classifications
 #'
@@ -46,12 +44,12 @@
 #' @importFrom Rdpack reprompt
 #'
 #' @export
-ExpNBFuzzyParam <- function(train, cl,  alphacut = 0.0001, metd = 2, alp = c(0.35, 0.7, 0.86), w = c(0.1,0.3,0.6), cores = 2) {
+ExpNBFuzzyParam <- function(train, cl,  alphacut = 0.0001, metd = 2, cores = 2) {
   UseMethod("ExpNBFuzzyParam")
 }
 
 #' @export
-ExpNBFuzzyParam.default <- function(train, cl,  alphacut = 0.0001, metd = 2, alp = c(0.35, 0.7, 0.86), w = c(0.1,0.3,0.6), cores = 2) {
+ExpNBFuzzyParam.default <- function(train, cl,  alphacut = 0.0001, metd = 2, cores = 2) {
 
   # --------------------------------------------------------
   # Estimating class parameters
@@ -61,13 +59,20 @@ ExpNBFuzzyParam.default <- function(train, cl,  alphacut = 0.0001, metd = 2, alp
   cols <- p_data$cols
   dados <- p_data$dados
   M <- p_data$M
+  MM <- as.numeric(M)
   # --------------------------------------------------------
   # --------------------------------------------------------
   # Finding lambdas for each class
-  lambdas <- lapply(1:length(unique(M)), function(i) colMeans(subset(dados, M == i)))
+  lambdas <- lapply(1:length(unique(M)), function(i) colMeans(subset(dados, MM == i)))
   # --------------------------------------------------------
   # --------------------------------------------------------
   # Estimating Triangular Parameters
+  # --------------------------------------------------------
+  if(metd > 3){ stop("metd argument must be <= 3.") }
+  # --------------------------------------------------------
+  alp=1
+  w=1
+  # --------------------------------------------------------
   alpha = alp
   if (metd != 4) {
     alpha <- seq(alphacut, 1.1, 0.1)
@@ -204,11 +209,6 @@ predict.ExpNBFuzzyParam <- function(object,
                     # Thorani Distance
                     Thoranidistance(vec_trian, M)
                     # ------------
-                  },
-                  "4" = {
-                    # ------------
-                    # Alpha-Order for a class of fuzzy sets
-                    AlphaOrderFuzzy(vec_trian, w, M)
                   }
     )
     # --------------------------------------------------------
