### 2023-08-16 #5613 @Jodavid
Hello @ajstewartlang, @MikeLydeamore, and @tianyi93,
returning with responses to suggestions

Below are each of them with responses regarding what has been accomplished.
I have added a checkmark and a response to each suggestion.

-----------------------------------------------------------

@MikeLydeamore:

Hi @ajstewartlang @tianyi93. I have completed my review of this work.
[x] The paper itself is well written, and bar one typo (exponencial distribution),
Response: It has been rectified.

[X] the only thing missing is a statement of other software solutions that exist for this problem.
Response: In the final paragraph of the introduction, information has been included regarding other libraries and software that encompass certain methods found within FuzzyClass. However, the majority of solutions, particularly those involving fuzzy probability, are exclusively available in FuzzyClass.

I believe there are some issues with the codebase and repository which should be amended before publication:

[X] Please fix the `pkgdown` site: https://leapigufpb.github.io/FuzzyClass/reference/index.html gives me a 404 error
Response: It has been rectified.

[X] Some automated testing of the software (beyond the current spellcheck) would be a great addition to robustness.
Response: The test_type.R file has been included as a classifiers test.

[X] Please add at least one vignette. This is a requirement for CRAN these days.
Response: A file containing an example of data analysis and instructions on how to utilize the classifier has been generated.

[X] Please add contribution guidelines to the README
Response: This section has been added to the README (https://github.com/leapigufpb/FuzzyClass/).

[X] Please add issue reporting guidelines to the README
Response: This section has been added to the README (https://github.com/leapigufpb/FuzzyClass/)

[X] The example in the README could do with much stronger documentation/commenting. It runs successfully but I don't know what it's supposed to do, or what the outputs mean.
Response: Steps regarding the code, problem statement, and usage example have been added.

[X] Dependencies: The `caret` library is used in the example, but this isn't included in the package install. It's a massive library, so I understand not wanting to include it as a dependency, but either an alternative `confusionMatrix` function should be sourced, or it needs to go into `DEPENDS`.
Response: Indeed, the caret package was employed for generating the confusion matrix. I also comprehend the rationale behind constructing an alternative confusion matrix. However, given the robustness of the confusionMatrix function in caret package, the authors opted to omit the caret-related portion from the example, instead opting for a simpler approach. Those desiring more resilient outcomes can avail themselves of the confusionMatrix function within caret.

Finally, the example uses a function `GauNBFuzzyParam`. The documentation for this has a few issues:

[X] Parameter `w` documentations talks about `metd == 4`, but the documentation about `metd` only says this can range between 1 and 3
Response: Given that this solution is still under development, the maintainers have chosen to exclude the metd=4 from the code.

[X] Argument `cores` isn't used in this function (passed through elsewhere), and this should be stated as I was looking for the parallel pool to spawn in this function, not elsewhere.
Response: The 'cores' argument is employed within the prediction function, owing to the computational costliness of these methods during prediction, rather than parameter estimation. A comment has been appended to the line concerning 'cores' within the manual of all functions, elucidating this detail.

@tianyi93:

 I have completed my review of this work. Few issues would suggest to amend before publication:

>Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

[X] Readme: please include running the command from R console
Response: The list of dependencies has been appended to the README; however, it is automatically installed during the package installation in R. The commands are present in .Rmd files, which are executed in R before generating the README. Moreover, an image of the R console displaying the installation and package loading commands has been incorporated.

>  Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

[X] No clear guidelines for third parties wishing to contribute to the software or seek support
Response: This section has been added to the README (https://github.com/leapigufpb/FuzzyClass/).

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

[X] Did not see much core functionality of the software documented
Response: The program's documentation/manual is available at the following link: https://cran.r-project.org/web/packages/FuzzyClass/FuzzyClass.pdf. Additionally, examples have been included in the README, and a Vignette has been created to provide usage examples.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?`

[X] Please add the above
Response: The test_type.R file has been included as a classifiers test.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

[X] there is no explicit description of how the "FuzzyClass" software compares to other commonly-used packages for classification tasks. The text primarily focuses on introducing the concept of Naive Bayes and Fuzzy Naive Bayes classifiers, their probability calculations, and the motivating example using the "FuzzyClass" package with real data.
Response:  In the final paragraph of the introduction, information has been included regarding other libraries and software that encompass certain methods found within FuzzyClass. However, the majority of solutions, particularly those involving fuzzy probability, are exclusively available in FuzzyClass.
