+++ vignettes/index.Rmd
@@ -4,7 +4,7 @@ author: "Jodavid Ferreira"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{FuzzyClass}
+  %\VignetteIndexEntry{FuzzyClassSolvingProblem}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
