+++ NAMESPACE
@@ -10,6 +10,7 @@ S3method(FuzzyGammaNaiveBayes,default)
 S3method(FuzzyGaussianNaiveBayes,default)
 S3method(FuzzyGeoNaiveBayes,default)
 S3method(FuzzyNaiveBayes,categorical)
+S3method(FuzzyNaiveBayes,continuo)
 S3method(FuzzyNaiveBayes,default)
 S3method(FuzzyPoissonNaiveBayes,default)
 S3method(FuzzyTrapezoidalNaiveBayes,default)
