### 2023-08-21 #5613 @ajstewartlang
> LGTM, no other major comments. Would be good if can make The program's documentation/manual easier for users to find. https://cran.r-project.org/web/packages/FuzzyClass/FuzzyClass.pdf.

@Jodavid might you be able to address this please?
