### 2023-08-22 #5613 @Jodavid
Hello @ajstewartlang, of course,

The link to the manual was already present in the paper in the last paragraph before the 'Acknowledgements'.

It has also been added to the readme at https://github.com/leapigufpb/FuzzyClass/ before the 'Dependencies' section with the following text:

"A package manual that showcases the existing classifiers and demonstrates how to use it can be found at the following link: https://cran.r-project.org/web/packages/FuzzyClass/FuzzyClass.pdf"

Thanks for suggestion @tianyi93
