### 2023-08-23 8c6edc4
update membership function
