+++ R/FuzzyGeoNaiveBayes.R
@@ -84,34 +84,36 @@ FuzzyGeoNaiveBayes.default <- function(train, cl, cores = 2, fuzzy = T) {
 
   # --------------------------------------------------------
   # --------------------------------------------------------
+  # Estimating Parameters
+  parametersC <- estimation_parameters_geo(M, cols, dados)
+  # --------------------------------------------------------
   # Estimating class memberships
-  pertinicesC <- lapply(1:length(unique(M)), function(i) {
-    lapply(1:cols, function(j) {
-      SubSet <- dados[M == unique(M)[i], j]
-      getMemberships(SubSet, intervalos)
-    })
-  })
+  Sturges <- Sturges(dados, M);
+  Comprim_Intervalo <- Comprim_Intervalo(dados, M, Sturges);
+  minimos <- minimos(dados, M, cols);
+  #Freq <- Freq(dados, M, Comprim_Intervalo, Sturges, minimos, cols);
+  #Pertinencia <- Pertinencia(Freq, dados, M);
+  MinimosDataFrame <- minomosdt_function(minimos, M, Comprim_Intervalo, Sturges, cols)
+  Frequencia <- Freq_esparsa(dados = dados,M = M, minomosdt = MinimosDataFrame, cols = cols)
+  Pertinencia <- Pertinencia_esparsa(M = M, Frequencia, cols = cols)
   # --------------------------------------------------------
+  # A priori probability of classes - considered equal
+  pk <- rep(1 / length(unique(M)), length(unique(M)))
   # --------------------------------------------------------
-  # Estimating Geometric Parameters
-  parametersC <- lapply(1:length(unique(M)), function(i) {
-    lapply(1:cols, function(j) {
-      SubSet <- dados[M == unique(M)[i], j]
-      param <- MASS::fitdistr(SubSet, "geometric")$estimate
-      return(param)
-    })
-  })
 
-  # --------------------------------------------------------
 
   # -------------------------------------------------------
   structure(list(
-    pertinicesC = pertinicesC,
     parametersC = parametersC,
+    minimos = minimos,
+    MinimosDataFrame = MinimosDataFrame,
     cols = cols,
     M = M,
     cores = cores,
-    intervalos = intervalos,
+    Comprim_Intervalo = Comprim_Intervalo,
+    Pertinencia = Pertinencia,
+    Sturges = Sturges,
+    pk = pk,
     fuzzy = fuzzy
   ),
   class = "FuzzyGeoNaiveBayes"
@@ -144,80 +146,32 @@ predict.FuzzyGeoNaiveBayes <- function(object,
   # --------------------------------------------------------
   test <- as.data.frame(newdata)
   # --------------------------------------------------------
-  pertinicesC <- object$pertinicesC
   parametersC <- object$parametersC
+  minimos <- object$minimos
+  MinimosDataFrame <-  object$MinimosDataFrame
   cols <- object$cols
   M <- object$M
   cores <- object$cores
-  intervalos <- object$intervalos
+  Comprim_Intervalo <- object$Comprim_Intervalo
+  Pertinencia <- object$Pertinencia
+  Sturges <- object$Sturges
+  pk <- object$pk
   fuzzy <- object$fuzzy
   # --------------------------------------------------------
 
   # --------------------------------------------------------
   # Classification
   # --------------
+  P <- density_values_geo(M, cols, test, parametersC, pk)
+  # --------------
   N_test <- nrow(test)
   # --------------
-  # Defining how many CPU cores to use
-  core <- parallel::makePSOCKcluster(cores)
-  doParallel::registerDoParallel(core)
-  # --------------
-  # loop start
-  R_M_obs <- foreach::foreach(h = 1:N_test, .combine = rbind) %dopar% {
-
-    # ------------
[…]
