+++ paper/refs.bib
@@ -19,7 +19,8 @@
 	title={Computational intelligence: principles, techniques and applications},
 	author={Konar, A.},
 	year={2006},
-	publisher={Springer Science \& Business Media}
+	publisher={Springer Science \& Business Media},
+	doi = {10.1093/comjnl/bxm073},
 }
 
 @book{efraim2011decision,
@@ -37,7 +38,8 @@
 	number={1-4},
 	pages={7--21},
 	year={2002},
-	publisher={Elsevier}
+	publisher={Elsevier},
+	doi = {10.1016/s0020-0255(02)00154-8},
 }
 
 @book{maulik2011computational,
@@ -45,7 +47,8 @@
 	author={Maulik, Ujjwal and Bandyopadhyay, Sanghamitra and Wang, Jason T},
 	volume={20},
 	year={2011},
-	publisher={Wiley Online Library}
+	publisher={Wiley Online Library},
+	doi = {10.1002/9780470872352.fmatter},
 }
 
 @inproceedings{pietramala2008genetic,
@@ -54,7 +57,8 @@
 	booktitle={Joint European Conference on Machine Learning and Knowledge Discovery in Databases},
 	pages={188--203},
 	year={2008},
-	organization={Springer}
+	organization={Springer},
+	doi = {10.1007/978-3-540-87481-2_13},
 }
 
 @article{vivekanandan2013intelligent,
@@ -103,7 +107,8 @@
 	volume={70},
 	pages={97--102},
 	year={2014},
-	publisher={Elsevier}
+	publisher={Elsevier},
+	doi = {10.1016/j.knosys.2014.05.006},
 }
 
 @inproceedings{de2014accuracy,
@@ -112,7 +117,8 @@
 	booktitle={Decision Making and Soft Computing: Proceedings of the 11th International FLINS Conference},
 	pages={161--166},
 	year={2014},
-	organization={World Scientific}
+	organization={World Scientific},
+	doi = {10.1142/9789814619998_0029},
 }
 
 @article{ferreira2015assessment,
@@ -149,7 +155,8 @@
 	number={Preprint},
 	pages={1--12},
 	year={2020},
-	publisher={IOS Press}
+	publisher={IOS Press},
+	doi = {10.3233/jifs-179431},
 }
 
 
@@ -176,7 +183,8 @@
   booktitle={2016 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE)},
   pages={745--750},
   year={2016},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/fuzz-ieee.2016.7737762 },
 }
 
 
@@ -191,7 +199,8 @@
   number={2},
   pages={117},
   year={2007},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi = {10.1088/0026-1394/44/2/003},
 }
 
 
@@ -211,6 +220,7 @@
     author = {Karline Soetaert and Peter M.J. Herman},
     pages = {372},
     note = {ISBN 978-1-4020-8623-6},
+    doi = {10.1016/s0075-9511(09)00026-7},
   }
   
   @inproceedings{de2020online,
@@ -219,7 +229,8 @@
   booktitle={Proc. FLINS},
   pages={446--454},
   year={2020},
-  organization={World Scientific}
+  organization={World Scientific},
+  doi = {10.1142/9789811223334_0054},
 }
 
 
@@ -231,7 +242,8 @@
   number={3},
   pages={849--861},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s40815-020-00936-4},
 }
 
 @article{zadeh1968,
@@ -272,7 +284,8 @@
   volume={4},
   number={1},
   pages={93--100},
[…]
