### 2023-08-23 #5613 @Jodavid
Hello @ajstewartlang,

I've checked all the DOIs, added the missing ones, and confirmed and corrected the existing ones.
The files are already updated in the repository.
