### 2023-08-24 #5613 @ajstewartlang
:wave; @Jodavid there still seem to be some DOIs missing (e.g., for this one: https://www.worldscientific.com/doi/abs/10.1142/9789812774118_0111) - they're not being picked up automatically - could you add those missing ones manually please?
