+++ paper/refs.bib
@@ -1,19 +1,64 @@
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% begin introduction
 
-@book{pathak2014beginning,
-	title={Beginning data science with R},
-	author={Pathak, MA.},
-	year={2014},
-	publisher={Springer}
+@incollection{de2006line,
+	title={On-line Training Evaluation in Virtual Reality Simulators using Fuzzy Bayes Rule},
+	author={Moraes, RM. and Machado, LS.},
+	booktitle={Applied Artificial Intelligence},
+	pages={791--798},
+	year={2006},
+	publisher={World Scientific},
+	doi = {10.1142/9789812774118_0111},
 }
 
-@book{webb2003statistical,
-	title={Statistical pattern recognition},
-	author={Webb, AR.},
-	year={2003},
-	publisher={John Wiley \& Sons}
+@incollection{de2010fuzzy,
+	title={Fuzzy gaussian naive bayes applied to online assessment in virtual reality simulators},
+	author={Moraes, RM. and Machado, LS.},
+	booktitle={Computational Intelligence: Foundations and Applications},
+	pages={243--248},
+	year={2010},
+	publisher={World Scientific},
+	doi = {10.1142/9789814324700_0035},
 }
+
+@inproceedings{de2018fuzzyGamma,
+	title={A fuzzy gamma naive bayes classifier},
+	author={{Moraes}, RM. and {Soares}, EAMG. and {Machado}, LS.},
+	booktitle={Data Science and Knowledge Engineering for Sensing Decision Support: Proceedings of the 13th International FLINS Conference (FLINS 2018)},
+	pages={691--699},
+	year={2018},
+	organization={World Scientific},
+	doi = {10.1142/9789813273238_0088},
+}
+
+@inproceedings{de2020new,
+	title={A new fuzzy beta naive Bayes classifier},
+	author={Moraes, RM. and Rodrigues, AKG. and Soares, EAMG. and Machado, LS.},
+	booktitle={Developments of Artificial Intelligence Technologies in Computation and Robotics: Proceedings of the 14th International FLINS Conference (FLINS 2020)},
+	pages={437--445},
+	year={2020},
+	organization={World Scientific},
+	doi = {10.1142/9789811223334_0053},
+}
+
+  
+@inproceedings{de2020online,
+	title={Online skills assessment in training based on virtual reality using a novel fuzzy triangular naive Bayes network},
+	author={Moraes, RM and Silva, ILA. and Machado, LS.},
+	booktitle={Proc. FLINS},
+	pages={446--454},
+	year={2020},
+	organization={World Scientific},
+	doi = {10.1142/9789811223334_0054},
+}
+
+
+@book{efraim2011decision,
+	title={Decision support and business intelligence systems},
+	author={Efraim, T.},
+	year={2011},
+	publisher={Pearson Education India},
+	doi = {10.1002/9780470634431 },
+}
+
 
 @book{konar2006computational,
 	title={Computational intelligence: principles, techniques and applications},
@@ -23,81 +68,15 @@
 	doi = {10.1093/comjnl/bxm073},
 }
 
-@book{efraim2011decision,
-	title={Decision support and business intelligence systems},
-	author={Efraim, T.},
-	year={2011},
-	publisher={Pearson Education India}
-}
 
-@article{de2002fuzzy,
-	title={Fuzzy expert systems architecture for image classification using mathematical morphology operators},
-	author={Moraes, RM. and Banon, GJF. and Sandri, SA.},
-	journal={Information Sciences},
[…]
