### 2023-08-24 #5613 @Jodavid
Hello @ajstewartlang,

I have reviewed the `.md` and `.bib` files and added the DOIs that were not appearing. I've updated the files in the repository. I believe they should now both be in the references
